Hoist ViewHolder into its own file
